requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
deep-translator==1.11.4
diff-match-patch==20241021
Django==5.1.3
django-import-export==4.3.3
idna==3.10
pillow==11.0.0
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.2
tablib==3.7.0
urllib3==2.2.3
